+++ NEWS.md
@@ -3,6 +3,7 @@
 ### Features
 * Set Authority DELETE job/action/field mapping profiles to display via Settings > Data import ([MODDICONV-436](https://folio-org.atlassian.net/browse/MODDICONV-436))
 * Add validation for DELETE JobProfiles ([MODDICONV-437](https://folio-org.atlassian.net/browse/MODDICONV-437))
+* Add DELETE Authority default mapping profile ([MODDICONV-447](https://folio-org.atlassian.net/browse/MODDICONV-447))
 
 ### Tech Dept
 * Delete default quickMarc profiles ([MODDICONV-445](https://folio-org.atlassian.net/browse/MODDICONV-445))